+++ app/(pages)/_data/navLinks.json
@@ -17,6 +17,6 @@
     },
     {
       "name": "Content Collection",
-      "slug": "/content-collection"
+      "slug": "/content/content"
     }
   ]
