+++ app/(pages)/content/[collection]/CollectionCard/CollectionCard.tsx
