+++ app/(pages)/content/[collection]/CollectionCard/MediaCard.tsx
