+++ app/(pages)/content/[collection]/CollectionCard/MediaPreview.module.scss
