+++ app/(pages)/content/[collection]/CollectionCard/MediaPreview.tsx
