+++ app/(pages)/content/[collection]/CollectionCard/PreviewCard.tsx
