+++ app/(pages)/content/[collection]/MediaGallery/MediaGallery.scss
@@ -1,0 +1,120 @@
+// mediagallery.scss
+@use 'media';
+
+.media-gallery-card {
+    display: flex;
+    flex-direction: row;
+    width: 80%;
+    margin: 0 auto; 
+    border: 1px solid #ccc;
+    border-radius: 8px;
+    margin-bottom: 20px;
+    padding: 20px; 
+    position: relative; 
+  }
+  
+  .media-gallery-card-number {
+    font-size: 24px; 
+    font-weight: bold;
+    color: rgba(0, 0, 0, 0.8); 
+    padding: 10px;
+    border-radius: 5px;
+    position: relative;
+    top: 50%;
+    transform: translateY(20%); 
+  }
+  
+  
+  .media-gallery-card-preview {
+    flex: 1;
+    max-width: 50%;
+    border-radius: 8px 0 0 8px;
+    position: relative;
+    top: 50%; 
+    transform: translateY(5%); 
+  }
+  
+  .media-gallery-card-preview img {
+    display: block;
+    max-width: 60%;
+    max-height: 60%; 
+    width: auto;
+    height: auto;
+    border-radius: 8px 0 0 8px;
+  }
+  
+  .media-gallery-card-details {
+    flex: 1;
+    margin-right: 50%; 
+  }
+  
+  .media-gallery-card-details h4 {
+    flex: 1;
+    padding: 20px;
+    padding-left: 0px; 
+    right: 200px;
+    
+  }
+  
+  @media (max-width: 768px) {
+    .media-gallery-card {
+      flex-direction: column;
+      width: 100%; 
+    }
+  
+    .media-gallery-card-preview {
+      max-width: 100%;
+      border-radius: 8px 8px 0 0;
+    }
+  }
+
+  .cover-image {
+    position: absolute;
+    top: 48%;
+    right: 45%; 
+    color: white;
+    background-color: rgb(117, 75, 185);
+    padding: 5px 10px;
+    border-radius: 5px;
+  }
+
+  .image-size {
+    position: absolute;
+    top: 50%; 
+    right: 30%; 
+    color: rgb(0, 0, 0);
+    padding: 5px 10px;
+    border-radius: 5px;
+  }
+
+  .replace-image-button {
+    position: absolute;
+    top: 50%; 
+    right: 20%;
+    background-color: #8175b7; 
+    color: white; 
+    padding: 5px 10px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+  }
+  
+  .replace-image-button:hover {
+    background-color: #9d9fc4; 
+  }
+
+  .delete-image-button {
+    position: absolute;
+    top: 45%;
+    right: 10%; 
+    background-color: #8175b7; 
+    color: white; 
+    padding: 5px 10px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+  }
+  
+  .delete-image-button:hover {
+    background-color: #9d9fc4; 
+  }
